+++ tests/integration/api/test_xai.py
@@ -48,12 +48,17 @@ def test_forward_explain(
     model_name = recipe_split[-1].split(".")[0]
     task = recipe_split[-2]
 
+<<<<<<< HEAD
     if "dino" in model_name:
         pytest.skip("DINO is not supported.")
 
     if "rtmdet_tiny" in recipe:
         # TODO (sungchul): enable xai for rtmdet_tiny (CVS-142651)
         pytest.skip("rtmdet_tiny on detection is not supported yet.")
+=======
+    if "dino" in model_name or "mobilenet_v4" in model_name or "rtmdet_tiny" in model_name:
+        pytest.skip(f"{model_name} is not supported.")
+>>>>>>> geti-classic
 
     engine = OTXEngine.from_config(
         config_path=recipe,
@@ -101,6 +106,7 @@ def test_predict_with_explain(
     model_name = recipe_split[-1].split(".")[0]
     task = recipe_split[-2]
 
+<<<<<<< HEAD
     if "dino" in model_name:
         pytest.skip("DINO is not supported.")
 
@@ -111,6 +117,10 @@ def test_predict_with_explain(
     if "yolox_tiny_tile" in recipe:
         # TODO (Galina): required to update model-api to 2.1
         pytest.skip("yolox_tiny_tile on detection requires model-api update")
+=======
+    if "dino" in model_name or "mobilenet_v4" in model_name or "rtmdet_tiny" in model_name:
+        pytest.skip(f"{model_name} is not supported.")
+>>>>>>> geti-classic
 
     tmp_path = tmp_path / f"otx_xai_{model_name}"
     engine = OTXEngine.from_config(